move to scan and respect pattern
